ai/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.24.0
azure-mgmt-core==1.6.0
azure-mgmt-postgresqlflexibleservers==1.1.0
blockbuster==1.5.25
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==44.0.3
distro==1.9.0
forbiddenfruit==0.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
langchain==0.3.27
langchain-core==0.3.74
langchain-mcp-adapters==0.1.9
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.4
langgraph-api==0.2.125
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.6
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.6.10
langgraph-sdk==0.2.0
langsmith==0.4.13
markdown-it-py==3.0.0
mcp==1.12.4
mdurl==0.1.2
msal==1.33.0
msal-extensions==1.3.1
openai==1.99.3
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sse-starlette==2.1.3
starlette==0.47.2
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.10.0
tqdm==4.67.1
truststore==0.10.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
xxhash==3.5.0
zstandard==0.23.0
